requirements.txt
Django==4.1.3
django-ninja==0.19.1
django-cors-headers==3.13.0
django-extensions==3.2.1
ipython==8.6.0
python-dotenv==0.21.0
psycopg2==2.9.5
PyJWT==2.6.0
cffi==1.15.1